fixed tox and created first test
